LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/fill-and-stroke-styles/2d.fillStyle.CSSHSL-expected.txt

2d.fillStyle.CSSHSL

CSSHSL works as color input


FAIL CSSHSL works as color input assert_approx_equals: Red channel of the pixel at (50, 25) expected 64 +/- 3 but got 0
